Begin process of web scraping
Because of the web page structure in the video and the web page
structure I'm actually using differ so much, it was not obvious how to
"follow along." I've made a start. To finish, I believe I would need to
do a bunch of "detective" work on my actual web page to determine how
to:

- Navigate the DOM tree to the "Sign in" button on the page
- Simulate checking that button. (This action then presents the
  credentials pop up.)
- Once the pop up is displayed, navigate the changed DOM to find the
  user name and password fields and enter the appropriate data.
- Once logged in, navigate to the "Account" page
- Finally, scrape the token used to access data.

All doable (and necessary if I needed to perform this activity in "real
life"), but the effort does not directly contribute to my goal of
learning to use `clj-http`.